+++ core/grafana/templates/flow_panel_template.json
