+++ core/layout/layout_manager.py
@@ -1,0 +1,6 @@
+from abc import ABC, abstractmethod
+
+class LayoutManager(ABC):
+    @abstractmethod
+    def apply(self, diagram, verbose=False):
+        pass
