+++ core/models/node.py
